requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pytest>=7.4.0
httpx>=0.25.0
python-dotenv>=1.0.0
playwright>=1.49.0
langchain-openai>=0.2.2
